Separate name into first_name and last_name; flag assoc cleanup
User.name was too ambiguous and often we needed to refer to just
first_name which was bug-prone. Now we allow user to set both
first_name and last_name directly, with a full_name helper to
combine them when needed.

Also fixed some mislabeled associations and the related tests.